% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_piesom.R
\name{plot.piesom.property}
\alias{plot.piesom.property}
\title{plot.piesom.property}
\usage{
\method{plot}{piesom.property}(object)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
plot.piesom.property
}
\examples{
data(iris)
piesom_setup(iris, Species ~ .) \%>\% piesom_som() \%>\% plot.piesom.property()
}
